Change name of SignUpPage to SignUpForm
